+++ avocado/core/safeloader.py
@@ -17,7 +17,11 @@ Safe (AST based) test loader module utilities
 """
 
 import ast
+import collections
+import imp
+import os
 import re
+import sys
 
 from ..utils import data_structures
 
@@ -154,3 +158,201 @@ def find_class_and_methods(path, method_pattern=None, base_class=None):
                 methods = data_structures.ordered_list_unique(methods)
             result[statement.name] = methods
     return result
+
+
+def get_methods_info(statement_body, class_tags):
+    """
+    Returns information on an Avocado instrumented test method
+    """
+    methods_info = []
+    for st in statement_body:
+        if (isinstance(st, ast.FunctionDef) and
+                st.name.startswith('test')):
+            docstring = ast.get_docstring(st)
+            mt_tags = get_docstring_directives_tags(docstring)
+            mt_tags.update(class_tags)
+
+            methods = [method for method, _ in methods_info]
+            if st.name not in methods:
+                methods_info.append((st.name, mt_tags))
+
+    return methods_info
+
+
+def find_avocado_tests(path, class_name=None):
+    """
+    Attempts to find Avocado instrumented tests from Python source files
+
+    :param path: path to a Python source code file
+    :type path: str
+    :param class_name: the specific class to be found
+    :type path: str
+    :returns: tuple where first item is dict with class name and additional
+              info such as method names and tags; the second item is
+              set of class names which look like avocado tests but are
+              force-disabled.
+    :rtype: tuple
+    """
+    # If only the Test class was imported from the avocado namespace
+    test_import = False
+    # The name used, in case of 'from avocado import Test as AvocadoTest'
+    test_import_name = None
+    # If the "avocado" module itself was imported
+    mod_import = False
+    # The name used, in case of 'import avocado as avocadolib'
+    mod_import_name = None
+    # The resulting test classes
+    result = collections.OrderedDict()
+    disabled = set()
+
+    if os.path.isdir(path):
+        path = os.path.join(path, "__init__.py")
+
+    with open(path) as source_file:
+        mod = ast.parse(source_file.read(), path)
+
+    for statement in mod.body:
+        # Looking for a 'from avocado import Test'
+        if (isinstance(statement, ast.ImportFrom) and
+                statement.module == 'avocado'):
+
+            for name in statement.names:
+                if name.name == 'Test':
+                    test_import = True
+                    if name.asname is not None:
+                        test_import_name = name.asname
+                    else:
+                        test_import_name = name.name
+                    break
+
+        # Looking for a 'import avocado'
+        elif isinstance(statement, ast.Import):
+            for name in statement.names:
+                if name.name == 'avocado':
+                    mod_import = True
+                    if name.asname is not None:
+                        mod_import_name = name.nasname
+                    else:
+                        mod_import_name = name.name
+
+        # Looking for a 'class Anything(anything):'
+        elif isinstance(statement, ast.ClassDef):
+
+            # class_name will exist only under recursion. In that
+            # case, we will only process the class if it has the
+            # expected class_name.
+            if class_name is not None and class_name != statement.name:
+                continue
+
+            docstring = ast.get_docstring(statement)
+            # Looking for a class that has in the docstring either
+            # ":avocado: enable" or ":avocado: disable
+            has_disable = check_docstring_directive(docstring, 'disable')
+            if (has_disable and class_name is None):
+                disabled.add(statement.name)
+                continue
+
+            cl_tags = get_docstring_directives_tags(docstring)
+
+            has_enable = check_docstring_directive(docstring, 'enable')
+            if (has_enable and class_name is None):
+                info = get_methods_info(statement.body, cl_tags)
+                result[statement.name] = info
+                continue
+
+            # Looking for the 'recursive' docstring or a 'class_name'
+            # (meaning we are under recursion)
+            has_recurse = check_docstring_directive(docstring, 'recursive')
+            if (has_recurse or class_name is not None):
+                info = get_methods_info(statement.body, cl_tags)
+                result[statement.name] = info
+
+                # Getting the list of parents of the current class
+                parents = statement.bases
+
+                # Searching the parents in the same module
+                for parent in parents[:]:
+                    # Looking for a 'class FooTest(module.Parent)'
+                    if isinstance(parent, ast.Attribute):
+                        parent_class = parent.attr
+                    # Looking for a 'class FooTest(Parent)'
+                    else:
+                        parent_class = parent.id
+                    res, dis = find_avocado_tests(path, parent_class)
+                    if res:
+                        parents.remove(parent)
+                        for cls in res:
+                            info.extend(res[cls])
+                    disabled.update(dis)
+
+                # If there are parents left to be discovered, they
+                # might be in a different module.
+                for parent in parents:
+                    if isinstance(parent, ast.Attribute):
+                        # Looking for a 'class FooTest(module.Parent)'
+                        parent_module = parent.value.id
+                        parent_class = parent.attr
+                    else:
+                        # Looking for a 'class FooTest(Parent)'
+                        parent_module = None
+                        parent_class = parent.id
+
+                    for node in mod.body:
+                        reference = None
+                        # Looking for 'from parent import class'
+                        if isinstance(node, ast.ImportFrom):
+                            reference = parent_class
+                        # Looking for 'import parent'
+                        elif isinstance(node, ast.Import):
+                            reference = parent_module
+
+                        if reference is None:
+                            continue
+
+                        for artifact in node.names:
+                            # Looking for a class alias
+                            # ('from parent import class as alias')
+                            if artifact.asname is not None:
+                                parent_class = reference = artifact.name
+                            # If the parent class or the parent module
+                            # is found in the imports, discover the
+                            # parent module path and find the parent
+                            # class there
+                            if artifact.name == reference:
+                                modules_paths = [os.path.dirname(path)]
+                                modules_paths.extend(sys.path)
+                                if parent_module is None:
+                                    parent_module = node.module
+                                _, ppath, _ = imp.find_module(parent_module,
+                                                              modules_paths)
+                                res, dis = find_avocado_tests(ppath,
+                                                              parent_class)
+                                if res:
+                                    for cls in res:
+                                        info.extend(res[cls])
+                                disabled.update(dis)
+
+                continue
+
+            if test_import:
+                base_ids = [base.id for base in statement.bases
+                            if hasattr(base, 'id')]
+                # Looking for a 'class FooTest(Test):'
+                if test_import_name in base_ids:
+                    info = get_methods_info(statement.body,
+                                            cl_tags)
+                    result[statement.name] = info
+                    continue
+
+            # Looking for a 'class FooTest(avocado.Test):'
+            if mod_import:
+                for base in statement.bases:
+                    module = base.value.id
+                    klass = base.attr
+                    if module == mod_import_name and klass == 'Test':
+                        info = get_methods_info(statement.body,
+                                                cl_tags)
+                        result[statement.name] = info
+                        continue
+
+    return result, disabled
