+++ selftests/unit/test_safeloader.py
@@ -1,9 +1,63 @@
 import ast
+import sys
+import os
 import re
 import unittest
 
 from avocado.core import safeloader
 from avocado.utils import script
+
+
+KEEP_METHODS_ORDER = '''
+from avocado import Test
+
+class MyClass(Test):
+    def test2(self):
+        pass
+
+    def testA(self):
+        pass
+
+    def test1(self):
+        pass
+
+    def testZZZ(self):
+        pass
+
+    def test(self):
+        pass
+'''
+
+RECURSIVE_DISCOVERY_TEST1 = """
+from avocado import Test
+
+class BaseClass(Test):
+    def test_basic(self):
+        pass
+
+class FirstChild(BaseClass):
+    def test_first_child(self):
+        pass
+
+class SecondChild(FirstChild):
+    '''
+    :avocado: disable
+    '''
+    def test_second_child(self):
+        pass
+"""
+
+RECURSIVE_DISCOVERY_TEST2 = """
+from avocado import Test
+from recursive_discovery_test1 import SecondChild
+
+class ThirdChild(Test, SecondChild):
+    '''
+    :avocado: recursive
+    '''
+    def test_third_child(self):
+        pass
+"""
 
 
 def get_this_file():
@@ -165,7 +219,9 @@ class FindClassAndMethods(UnlimitedDiff):
             'FindClassAndMethods': ['test_self',
                                     'test_with_pattern',
                                     'test_with_base_class',
-                                    'test_with_pattern_and_base_class'],
+                                    'test_with_pattern_and_base_class',
+                                    'test_methods_order',
+                                    'test_recursive_discovery'],
             'UnlimitedDiff': ['setUp']
         }
         found = safeloader.find_class_and_methods(get_this_file())
@@ -187,7 +243,9 @@ class FindClassAndMethods(UnlimitedDiff):
             'FindClassAndMethods': ['test_self',
                                     'test_with_pattern',
                                     'test_with_base_class',
-                                    'test_with_pattern_and_base_class'],
+                                    'test_with_pattern_and_base_class',
+                                    'test_methods_order',
+                                    'test_recursive_discovery'],
             'UnlimitedDiff': []
         }
         found = safeloader.find_class_and_methods(get_this_file(),
@@ -199,7 +257,9 @@ class FindClassAndMethods(UnlimitedDiff):
             'FindClassAndMethods': ['test_self',
                                     'test_with_pattern',
                                     'test_with_base_class',
-                                    'test_with_pattern_and_base_class'],
+                                    'test_with_pattern_and_base_class',
+                                    'test_methods_order',
+                                    'test_recursive_discovery'],
         }
         found = safeloader.find_class_and_methods(get_this_file(),
                                                   base_class='UnlimitedDiff')
@@ -209,12 +269,41 @@ class FindClassAndMethods(UnlimitedDiff):
         reference = {
             'FindClassAndMethods': ['test_with_pattern',
                                     'test_with_base_class',
-                                    'test_with_pattern_and_base_class'],
+                                    'test_with_pattern_and_base_class']
         }
         found = safeloader.find_class_and_methods(get_this_file(),
                                                   re.compile(r'test_with.*'),
                                                   'UnlimitedDiff')
         self.assertEqual(reference, found)
+
+    def test_methods_order(self):
+        avocado_keep_methods_order = script.TemporaryScript(
+            'keepmethodsorder.py',
+            KEEP_METHODS_ORDER)
+        avocado_keep_methods_order.save()
+        expected_order = ['test2', 'testA', 'test1', 'testZZZ', 'test']
+        tests = safeloader.find_avocado_tests(avocado_keep_methods_order.path)[0]
+        methods = [method[0] for method in tests['MyClass']]
+        self.assertEqual(expected_order, methods)
+        avocado_keep_methods_order.remove()
+
+    def test_recursive_discovery(self):
+        avocado_recursive_discovery_test1 = script.TemporaryScript(
+            'recursive_discovery_test1.py',
+            RECURSIVE_DISCOVERY_TEST1)
+        avocado_recursive_discovery_test1.save()
+        avocado_recursive_discovery_test2 = script.TemporaryScript(
+            'recursive_discovery_test2.py',
+            RECURSIVE_DISCOVERY_TEST2)
+        avocado_recursive_discovery_test2.save()
+
+        sys.path.append(os.path.dirname(avocado_recursive_discovery_test1.path))
+        tests = safeloader.find_avocado_tests(avocado_recursive_discovery_test2.path)[0]
+        expected = {'ThirdChild': [('test_third_child', set([])),
+                                   ('test_second_child', set([])),
+                                   ('test_first_child', set([])),
+                                   ('test_basic', set([]))]}
+        self.assertEqual(expected, tests)
 
 
 if __name__ == '__main__':
